new map structure
as the name suggests
